implement headline and close button into default modal
